+++ lib/common/common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cpu;
 pub mod defaults;
+pub mod disk;
 pub mod fixed_length_priority_queue;
 pub mod math;
 pub mod panic;
